requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
gTTS==1.2.0
gTTS-token==1.1.1
idna==2.5
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.6
lxml==3.8.0
MarkupSafe==1.0
newspaper3k==0.2.2
nltk==3.2.4
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.1
requests-file==1.4.2
six==1.10.0
tldextract==2.1.0
urllib3==1.21.1
Werkzeug==0.12.2
